This dataset is a subset of the 1987 National Indonesia Contraceptive Prevalence Survey created by Tjen-Sien Lim (limt@stat.wisc.edu) and donated to UCI Machine Learning Repository by him on June 7, 1997 (Dua & Graff, 2019). The samples are married women who were either not pregnant or do not know if they were at the time of interview. The problem is to predict the current contraceptive method choice (no use, long-term methods, or short-term methods) of a woman based on her demographic and socio-economic characteristics. Machine learning models can be used for two purposes: understanding and predicting. In this case, we can both understand the underlying behavior of Indonesians making decisions about contraceptive methods as well as predict which choice a given Indonesian woman would make.
The ultimate goal of the Family Planning Program is to educate women about contraceptive use so that they are empowered to make their own decisions about contraception and family planning. Using a decision tree model allows us to understand what the biggest factors are in the behavior patterns of these women. For example, looking at the visualization of the pairings above, it is clear that young women who are not well educated are extremely likely to be using no contraception, while older women, with more kids and higher standard of living, are likely to be using contraception methods. The Indonesian government should target young women with educational campaigns about contraceptive methods in secondary schools and universities.